+++ ir/opt/opt_osr.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1995-2008 University of Karlsruhe.  All right reserved.
+ * Copyright (C) 1995-2011 University of Karlsruhe.  All right reserved.
  *
  * This file is part of libFirm.
  *
@@ -79,7 +79,7 @@ typedef struct node_entry {
 typedef struct iv_env {
 	struct obstack obst;    /**< an obstack for allocations */
 	ir_node  **stack;       /**< the node stack */
-	int      tos;           /**< tos index */
+	size_t   tos;           /**< tos index */
 	unsigned nextDFSnum;    /**< the current DFS number */
 	unsigned POnum;         /**< current post order number */
 	set      *quad_map;     /**< a map from (op, iv, rc) to node */
@@ -950,7 +950,7 @@ static void push(iv_env *env, ir_node *n)
 	node_entry *e;
 
 	if (env->tos == ARR_LEN(env->stack)) {
-		int nlen = ARR_LEN(env->stack) * 2;
+		size_t nlen = ARR_LEN(env->stack) * 2;
 		ARR_RESIZE(ir_node *, env->stack, nlen);
 	}
 	env->stack[env->tos++] = n;
